requirements.txt
tornado==6.1

websocket-client~=0.57.0
mysqlclient~=2.0.1